+++ DemoApps/GRDBDemoiOS7/GRDBDemoiOS7/Person.swift
@@ -14,6 +14,10 @@ class Person : Record {
     
     // MARK: - Record
     
+    required init(row: Row) {
+        super.init(row: row)
+    }
+    
     override class func databaseTableName() -> String {
         return "persons"
     }
@@ -30,9 +34,5 @@ class Person : Record {
             "id": id,
             "firstName": firstName,
             "lastName": lastName]
-    }
-
-    required init(row: Row) {
-        super.init(row: row)
     }
 }