+++ GRDB/Foundation/NSURL.swift
@@ -19,4 +19,12 @@ extension NSURL : DatabaseValueConvertible {
         }
         return self.init(string: string)
     }
+    
+    public static func fromRow(row: Row) -> Self {
+        // TOOD: test
+        guard let url = fromDatabaseValue(row.databaseValues.first!) else {
+            fatalError("Could not convert \(row.databaseValues.first!) to NSURL.")
+        }
+        return url
+    }
 }